+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - NFS drive automatically attached when using cirrascale clusters.
 - Added `--save-spec PATH` option to `beaker run` for saving the generated experiment spec to a YAML file.
+- Added automatic upgrade checks. You'll get a warning message if your installation of gantry is out-of-date.
 
 ## [v0.2.1](https://github.com/allenai/beaker-gantry/releases/tag/v0.2.1) - 2022-05-16